+++ HISTORY.md
@@ -4,7 +4,7 @@ Release History
 ***********************************************************************************************************************
 # TODO
 1. add TIMEOUT (use start in thread!) for print! use timeout for GETATTR!!!
-2. realise CHANGE_state/COMPARE_objects (one from different states like thread before and after start)!
+2. realise PRINT_DIFFS=CHANGE_state/COMPARE_objects (one from different states like thread before and after start)!
    - this is about to save object STATE!
    - add parameter show only diffs or show all
    - add TESTS after this step!